test: share component registration fixture
